Add files via upload
this is the python file that I've done while being bored on lectures about machine learning. This game is literally why I've passed that module with a first. I'm more of a practice than theory person and so while programming this I was able to actually pay attention haha :D